+++ Cargo.toml
@@ -7,7 +7,7 @@ edition = "2021"
 
 [dependencies]
 tokio = { version = "1.29.1", features = ["rt", "macros", "io-std"] }
-tower-lsp = "0.19.0"
+tower-lsp = "0.20.0"
 async-trait = "0.1"
 
 semver = { version = "1", features = ["serde"] }
